breaks components into separate folders
