+++ src/containers/SurveyCardData/SurveyCardData.scss
