+++ src/index.scss
@@ -9,7 +9,7 @@
 @import './components/Loader/Loader.scss';
 @import './containers/RecipientForm/RecipientForm.scss';
 @import './containers/SurveyCard/SurveyCard.scss';
-@import './containers/SurveyCard/SurveyCardData.scss';
+@import './containers/SurveyCardData/SurveyCardData.scss';
 @import './containers/StudentSurvey/StudentSurvey.scss';
 
 body {
