+++ src/utils/AverageResponses.js
